+++ documents/index.js
@@ -1,0 +1,4 @@
+export { default as DocumentsFolderList } from './DocumentsFolderList.vue'
+export { default as DocumentViewerModal } from './DocumentViewerModal.vue'
+export { default as DocumentOpeningOverlay } from './DocumentOpeningOverlay.vue'
+export * from './sharePointHost.js'